This commit was generated by cvs2svn to compensate for changes in r12179, which included commits to RCS files with non-trunk default branches.
git-svn-id: https://svn.code.sf.net/p/vice-emu/code/trunk@12180 379a1393-f5fb-40a0-bcee-ef074d9b53f7

+++ vice/src/vic20/vic-draw.c
@@ -134,7 +134,8 @@ inline static void draw(BYTE *p, unsigned int xs, unsigned int xe, int reverse,
     c[1] = VIC_PIXEL(vic.mc_border_color);
     c[3] = VIC_PIXEL(vic.auxiliary_color);
 
-    for (i = xs; i <= xe; i++, p += 8 * VIC_PIXEL_WIDTH) {
+    /* xe can be -1.  */
+    for (i = xs; (int)i <= (int)xe; i++, p += 8 * VIC_PIXEL_WIDTH) {
         b = *(vic.color_ptr + vic.memptr + i);
         c[2] = VIC_PIXEL(b & 0x7);
         if (reverse & !(b & 0x8))
